School MGT

+++ SchoolManagementSystem/SchoolManagementSystem/Controllers/StudentTablesController.cs
@@ -13,7 +13,7 @@ namespace SchoolManagementSystem.Controllers
     public class StudentTablesController : Controller
     {
         private SchoolMgtSysDbEntities db = new SchoolMgtSysDbEntities();
-
+       
         // GET: StudentTables
         public ActionResult Index()
         {
@@ -21,7 +21,7 @@ namespace SchoolManagementSystem.Controllers
             {
                 return RedirectToAction("Login", "Home");
             }
-            var studentTables = db.StudentTables.Include(s => s.ProgrameTable).Include(s => s.SessionTable).Include(s => s.UserTable).OrderByDescending(s => s.StudentID);
+            var studentTables = db.StudentTables.Include(s => s.ProgrameTable).Include(s => s.SessionTable).Include(s => s.UserTable).Include(s => s.ClassTable);
             return View(studentTables.ToList());
         }
 
@@ -54,6 +54,7 @@ namespace SchoolManagementSystem.Controllers
             ViewBag.ProgrameID = new SelectList(db.ProgrameTables, "ProgrameID", "Name");
             ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name");
             ViewBag.UserID = new SelectList(db.UserTables, "UserID", "FullName");
+            ViewBag.ClassID = new SelectList(db.ClassTables, "ClassID", "Name");
             return View();
         }
 
@@ -71,11 +72,23 @@ namespace SchoolManagementSystem.Controllers
             int userid = Convert.ToInt32(Convert.ToString(Session["UserID"]));
             studentTable.UserID = userid;
             studentTable.Photo = "/Content/StudentPhoto/default.png";
+
             if (ModelState.IsValid)
             {
+                var user = new UserTable();
+                user.Address = studentTable.Address;
+                user.ContactNo = studentTable.ContactNo;
+                user.EmailAddress = studentTable.EmailAddress;
+                user.FullName = studentTable.Name;
+                user.UserName = studentTable.Name;
+                user.UserTypeID = 3;
+                user.Password = "123456";
+                db.UserTables.Add(user);
+                studentTable.UserID = user.UserID;
+
                 db.StudentTables.Add(studentTable);
                 db.SaveChanges();
-                if (studentTable.Photo != null)
+                if (studentTable.PhotoFile != null)
                 {
                     var folder = "/Content/StudentPhoto";
                     var file = string.Format("{0}.png", studentTable.StudentID);
@@ -88,13 +101,13 @@ namespace SchoolManagementSystem.Controllers
                         db.SaveChanges();
                     }
                 }
-
                 return RedirectToAction("Index");
             }
 
             ViewBag.ProgrameID = new SelectList(db.ProgrameTables, "ProgrameID", "Name", studentTable.ProgrameID);
             ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name", studentTable.SessionID);
             ViewBag.UserID = new SelectList(db.UserTables, "UserID", "FullName", studentTable.UserID);
+            ViewBag.ClassID = new SelectList(db.ClassTables, "ClassID", "Name", studentTable.ClassID);
             return View(studentTable);
         }
 
@@ -117,6 +130,7 @@ namespace SchoolManagementSystem.Controllers
             ViewBag.ProgrameID = new SelectList(db.ProgrameTables, "ProgrameID", "Name", studentTable.ProgrameID);
             ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name", studentTable.SessionID);
             ViewBag.UserID = new SelectList(db.UserTables, "UserID", "FullName", studentTable.UserID);
+            ViewBag.ClassID = new SelectList(db.ClassTables, "ClassID", "Name", studentTable.ClassID);
             return View(studentTable);
         }
 
@@ -133,20 +147,17 @@ namespace SchoolManagementSystem.Controllers
             }
             int userid = Convert.ToInt32(Convert.ToString(Session["UserID"]));
             studentTable.UserID = userid;
-            studentTable.Photo = "/Content/StudentPhoto/default.png";
             if (ModelState.IsValid)
             {
-                if (studentTable.Photo != null)
+                var folder = "/Content/StudentPhoto";
+                var file = string.Format("{0}.png", studentTable.StudentID);
+                var response = FileHelper.UploadFile.UploadPhoto(studentTable.PhotoFile, folder, file);
+                if (response)
                 {
-                    var folder = "/Content/StudentPhoto";
-                    var file = string.Format("{0}.png", studentTable.StudentID);
-                    var response = FileHelper.UploadFile.UploadPhoto(studentTable.PhotoFile, folder, file);
-                    if (response)
-                    {
-                        var pic = string.Format("{0}/{1}", folder, file);
-                        studentTable.Photo = pic;
-                    }
+                    var pic = string.Format("{0}/{1}", folder, file);
+                    studentTable.Photo = pic;
                 }
+
                 db.Entry(studentTable).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -154,12 +165,17 @@ namespace SchoolManagementSystem.Controllers
             ViewBag.ProgrameID = new SelectList(db.ProgrameTables, "ProgrameID", "Name", studentTable.ProgrameID);
             ViewBag.SessionID = new SelectList(db.SessionTables, "SessionID", "Name", studentTable.SessionID);
             ViewBag.UserID = new SelectList(db.UserTables, "UserID", "FullName", studentTable.UserID);
+            ViewBag.ClassID = new SelectList(db.ClassTables, "ClassID", "Name", studentTable.ClassID);
             return View(studentTable);
         }
 
         // GET: StudentTables/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -177,6 +193,10 @@ namespace SchoolManagementSystem.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (string.IsNullOrEmpty(Convert.ToString(Session["UserName"])))
+            {
+                return RedirectToAction("Login", "Home");
+            }
             StudentTable studentTable = db.StudentTables.Find(id);
             db.StudentTables.Remove(studentTable);
             db.SaveChanges();
